Fix data download CSV button on Firefox
The download CSV button on the "Explore the Data" page works by creating
a hidden link with the CSV data and triggering a click event on it. That
was working fine on Chrome and Safari, but we noticed it wasn't working
for Firefox. As it turns out, for this tactic to work on Firefox, the
hidden link needs to actually be included somewhere in the DOM.

This change updates the download CSV button by adding the hidden link to
the DOM so that it will work on Firefox.

https://stackoverflow.com/a/40676268
closes https://github.com/texas-justice-initiative/website-nextjs/issues/101