Fix menu toggling logic
There was a bug in the existing code that caused the menu toggle code to
run whenever any link in the header was clicked.

e.g. when navigating to explore the data one lost the ability to scroll,
because the Page component thought the menu was open and was therefore
using `overflow: hidden`