Add new newsfeed stories to top of list
Currently, when we add a new newsfeed story in the CMS, the new story is
added to the bottom of the list. Since we generally want new stories to
appear at the top of the feed, that means we need to drag the new story
to the top of the list, which is becoming more tedious as we add new stories.

This change takes advantage of the new `add_to_top` configuration option
for lists in Netlify CMS (see https://github.com/netlify/netlify-cms/pull/4465)
to have new newsfeed stories added to the top of the list.

Closes https://github.com/texas-justice-initiative/website-nextjs/issues/428